Chore: Renamed cloister_stage.py to cloister.py and test_cloister_stage.py to test_cloister.py
